pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenthalt"
version = "0.1.0"
description = "Production-grade guardrails for AI agent function calls. Prevent overspending, unauthorized purchases, unsafe deletions, and more — with Human-in-the-Loop approval."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "HZYAI", email = "hello@hzy.ai"},
]
keywords = ["ai", "agents", "guardrails", "safety", "human-in-the-loop", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
langchain = ["langchain-core>=0.1"]
dashboard = ["fastapi>=0.100", "uvicorn>=0.20", "websockets>=11.0"]
all = ["openai>=1.0", "langchain-core>=0.1", "fastapi>=0.100", "uvicorn>=0.20", "websockets>=11.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/hzyai/agenthalt"
Documentation = "https://github.com/hzyai/agenthalt#readme"
Repository = "https://github.com/hzyai/agenthalt"
Issues = "https://github.com/hzyai/agenthalt/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agenthalt"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
